minor refactoring of Lambda API for better extensibility (#5133)
